refactor(ios): split workspace session presentation
